requirements.txt
openai==0.28.1
langchain==0.0.350
python-dotenv==1.0.0
markdown==3.5.1